requirements.txt
1337x==1.2.5
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
exceptiongroup==1.2.1
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
platformdirs==4.2.1
pyparsing==3.1.2
requests==2.31.0
requests-cache==1.2.0
requests-toolbelt==1.0.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.11.0
url-normalize==1.4.3
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.18.2
